requirements.txt
python-dotenv
atlassian-python-api
jira
openai
requests
flask
flask-bootstrap 